Add userService and its controller and refactor authentication

+++ src/main/kotlin/org/library/infrastructure/HibernateUserRepository.kt
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository
 import java.util.*
 
 @Repository
-interface HibernateUserRepository : UserRepository, CrudRepository<User?, Int?> {
+interface HibernateUserRepository : UserRepository, CrudRepository<User?, Long?> {
     override fun findByUsername(username: String?): Optional<User?>?
 }
